Added some tests
