+++ package.json
@@ -9,7 +9,7 @@
     "test-overview": "jest test/overview/test.js",
     "test-questions": "jest question",
     "test-reviews": "jest test/reviews/test.js",
-    "test-related": "jest test/related-products/test.js",
+    "test-related": "jest related-products",
     "server-dev": "nodemon server/index.js",
     "react-dev": "webpack -w"
   },
@@ -59,6 +59,7 @@
     "jquery": "^3.6.0",
     "nodemon": "^2.0.15",
     "redux-devtools-extension": "^2.13.9",
+    "sinon": "^13.0.1",
     "terser-webpack-plugin": "^5.3.1",
     "underscore": "^1.13.2",
     "webpack": "^5.68.0",
